Moved icons to shared
